refactor: 모달 관련 컴포넌트 로직 useTimerPageModal로 이동
